[qalab] ready to qa ServerDriverSelenium
+ more vars on settings
+ updated ServerDriverSelenium
+ starting ServerDriverAppium
+ removed old code from script

+++ qalab/core/server_driver_base.py
@@ -3,8 +3,11 @@
 """TODO: doc module"""
 
 
-import qalab.configs.settings as SETTINGS
+import os
 import shutil
+import subprocess
+import wget
+import qalab.configs.settings as SETTINGS
 
 
 class ServerDriverBase(object):
@@ -16,15 +19,18 @@ class ServerDriverBase(object):
     _server_driver = 'base'
 
 
-    def __init__(self, mode):
+    def __init__(self, logger, mode):
         """TODO: doc method"""
+        if logger is None:
+            raise Exception("param 'logger' can't be None")
+        self.logger = logger
         if mode is None:
             raise Exception("param 'mode' can't be None")
         if mode not in ['hub', 'node']:
             raise Exception('Select valid mode, values are: [hub, node]')
         self._mode = mode
-        _config_path_example = SETTINGS.DRIVER_CONFIG_PATH_EXAMPLE
-        _config_path = SETTINGS.DRIVER_CONFIG_PATH
+        self._config_path_example = SETTINGS.DRIVER_CONFIG_PATH_EXAMPLE
+        self._config_path = SETTINGS.DRIVER_CONFIG_PATH
 
     def _configure(self, server_driver):
         """
@@ -36,18 +42,43 @@ class ServerDriverBase(object):
             raise Exception(
                 'Select valid mode, values are: [selenium, selendroid, appium]'
             )
+        self.logger.info('Configuring server_driver={}'.format(server_driver))
         self._config_path_example = self._config_path_example.format(
             SETTINGS.PATH_CONFIG, server_driver, self._mode)
         self._config_path = self._config_path.format(
             SETTINGS.PATH_CONFIG, server_driver, self._mode)
-        shutil.copy2(
-            self._config_path_example,
-            self._config_path)
+        if os.path.exists(self._config_path):
+            self.logger.warning('File exist, not copying from example')
+        else:
+            shutil.copy2(
+                self._config_path_example,
+                self._config_path)
 
     def install(self):
         """Installation process"""
-        raise NotImplementedError('need to be overrided on inherit classes')
+        raise NotImplementedError(
+            'need to be overrided on inherit classes')
 
     def start(self):
         """Installation process"""
-        raise NotImplementedError('need to be overrided on inherit classes')
+        raise NotImplementedError(
+            'need to be overrided on inherit classes')
+
+    def download_file(self, url, out_path):
+        """
+        Download file from 'url' to 'out_path'
+         logging process start and end
+        """
+        self.logger.info('Downloading file: ...')
+        self.logger.debug('  url={}'.format(url))
+        self.logger.debug('  out_path={}'.format(out_path))
+        wget.download(url, out=out_path)
+        self.logger.info('Downloading file: DONE')
+
+    def command_exec(self, cmd_args):
+        try:
+            logger.info("Executing command : {}".format(cmd_args))
+            return subprocess.call(cmd_args)
+        except Exception as err:
+            raise Exception(
+                "Failed at handle command: {}".format(cmd_args))
